# docs/swagger.yaml
basePath: /api
definitions:
  presenter.FileObject:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  presenter.ResponseHTTP:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: hieubq90@gmail.com
    name: Bùi Quang Hiếu
  description: This is an API for File Server Application
  title: File Management Service
  version: "1.0"
paths:
  /{project}/{folder}/{filename}:
    get:
      consumes:
      - multipart/form-data
      description: Allow download
      parameters:
      - description: Project Name
        in: path
        name: project
        required: true
        type: string
      - description: Folder Name
        in: path
        name: folder
        required: true
        type: string
      - description: File Name
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "301":
          description: Moved Permanently
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.ResponseHTTP'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/presenter.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/presenter.ResponseHTTP'
      summary: Download file
      tags:
      - files
  /{project}/{folder}/files:
    post:
      consumes:
      - multipart/form-data
      description: Allow upload single or multiple file
      parameters:
      - description: Project Name
        in: path
        name: project
        required: true
        type: string
      - description: Folder Name
        in: path
        name: folder
        required: true
        type: string
      - description: Files
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/presenter.ResponseHTTP'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/presenter.FileObject'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/presenter.ResponseHTTP'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/presenter.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/presenter.ResponseHTTP'
      summary: Upload file
      tags:
      - files
swagger: "2.0"